+++ CHANGELOG.rst
@@ -10,7 +10,11 @@ this scheme are considered to be bugs.
 `Unreleased`_
 -------------
 
-- Nothing yet.
+Added
+~~~~~
+
+- Added support for creating plans.
+- Added support for querying singular and lists of plans.
 
 `0.2.0`_ - 2015-09-18
 ---------------------